+++ tests/testthat/ollama/localhost-11434/api/chat-1e1974-POST.R
@@ -1,5 +1,5 @@
 structure(list(method = "POST", url = "http://localhost:11434/api/chat", 
     status_code = 200L, headers = structure(list(`Content-Type` = "application/json; charset=utf-8", 
-        Date = "Thu, 20 Mar 2025 10:23:59 GMT", `Content-Length` = "328"), class = "httr2_headers"), 
-    body = charToRaw("{\"model\":\"gemma2\",\"created_at\":\"2025-03-20T10:23:59.608578Z\",\"message\":{\"role\":\"assistant\",\"content\":\"Hello! 👋  How can I help you today?\"},\"done_reason\":\"stop\",\"done\":true,\"total_duration\":2775177667,\"load_duration\":70621583,\"prompt_eval_count\":26,\"prompt_eval_duration\":2240000000,\"eval_count\":12,\"eval_duration\":462000000}"), 
+        Date = "Thu, 21 Aug 2025 12:41:21 GMT", `Content-Length` = "330"), redact = character(0), class = "httr2_headers"), 
+    body = charToRaw("{\"model\":\"gemma2\",\"created_at\":\"2025-08-21T12:41:21.294827Z\",\"message\":{\"role\":\"assistant\",\"content\":\"Hello! 👋  How can I help you today?\"},\"done_reason\":\"stop\",\"done\":true,\"total_duration\":9819624917,\"load_duration\":4533008625,\"prompt_eval_count\":26,\"prompt_eval_duration\":4819809375,\"eval_count\":12,\"eval_duration\":462770333}"), 
     cache = new.env(parent = emptyenv())), class = "httr2_response")